+++ src/Box_TEST.cc
@@ -49,8 +49,10 @@ TEST(BoxTest, Inherit)
     EXPECT_TRUE(box != NULL);
   }
 
-  EXPECT_TRUE(box->Min() == math::Vector3d(0, 0, 0));
-  EXPECT_TRUE(box->Max() == math::Vector3d(0, 0, 0));
+  EXPECT_TRUE(box->Min() == math::Vector3d(math::MAX_D, math::MAX_D,
+      math::MAX_D));
+  EXPECT_TRUE(box->Max() == math::Vector3d(math::LOW_D, math::LOW_D,
+      math::LOW_D));
 
   {
     delete box;
@@ -68,8 +70,10 @@ TEST(BoxTest, EmptyConstructorNew)
     EXPECT_TRUE(box != NULL);
   }
 
-  EXPECT_TRUE(box->Min() == math::Vector3d(0, 0, 0));
-  EXPECT_TRUE(box->Max() == math::Vector3d(0, 0, 0));
+  EXPECT_TRUE(box->Min() == math::Vector3d(math::MAX_D, math::MAX_D,
+      math::MAX_D));
+  EXPECT_TRUE(box->Max() == math::Vector3d(math::LOW_D, math::LOW_D,
+      math::LOW_D));
 
   {
     delete box;
@@ -81,8 +85,10 @@ TEST(BoxTest, EmptyConstructorNew)
 TEST(BoxTest, EmptyConstructor)
 {
   math::Box box;
-  EXPECT_TRUE(box.Min() == math::Vector3d(0, 0, 0));
-  EXPECT_TRUE(box.Max() == math::Vector3d(0, 0, 0));
+  EXPECT_TRUE(box.Min() == math::Vector3d(math::MAX_D, math::MAX_D,
+      math::MAX_D));
+  EXPECT_TRUE(box.Max() == math::Vector3d(math::LOW_D, math::LOW_D,
+      math::LOW_D));
 }
 
 /////////////////////////////////////////////////
@@ -98,6 +104,35 @@ TEST_F(ExampleBox, CopyConstructor)
   math::Box box1(box);
   EXPECT_TRUE(box1.Min() == box.Min());
   EXPECT_TRUE(box1.Max() == box.Max());
+}
+
+/////////////////////////////////////////////////
+TEST_F(ExampleBox, ManuallySet)
+{
+  math::Box box1;
+  box1.Min().Set(-2, 2, 3);
+  box1.Max().Set(0, 0, 3);
+
+  box1 += box;
+  EXPECT_DOUBLE_EQ(box1.Min().X(), -2);
+  EXPECT_DOUBLE_EQ(box1.Min().Y(), -2);
+  EXPECT_DOUBLE_EQ(box1.Min().Z(),  2);
+
+  EXPECT_DOUBLE_EQ(box1.Max().X(), 1);
+  EXPECT_DOUBLE_EQ(box1.Max().Y(), 0);
+  EXPECT_DOUBLE_EQ(box1.Max().Z(), 3);
+
+  math::Box box2;
+  box2.Max().Set(1, 1, 1);
+  EXPECT_DOUBLE_EQ(box2.Size().X(), 0);
+  EXPECT_DOUBLE_EQ(box2.Size().Y(), 0);
+  EXPECT_DOUBLE_EQ(box2.Size().Z(), 0);
+
+  math::Box box3;
+  box3.Min().Set(-1, -1, -1);
+  EXPECT_DOUBLE_EQ(box3.Size().X(), 0);
+  EXPECT_DOUBLE_EQ(box3.Size().Y(), 0);
+  EXPECT_DOUBLE_EQ(box3.Size().Z(), 0);
 }
 
 /////////////////////////////////////////////////
@@ -127,13 +162,38 @@ TEST(BoxTest, MergeEmpty)
   math::Box box2;
 
   box1.Merge(box2);
-  EXPECT_NEAR(box1.Min().X(), 0, 1e-6);
-  EXPECT_NEAR(box1.Min().Y(), 0, 1e-6);
-  EXPECT_NEAR(box1.Min().Z(), 0, 1e-6);
+  EXPECT_DOUBLE_EQ(box1.Min().X(), math::MAX_D);
+  EXPECT_DOUBLE_EQ(box1.Min().Y(), math::MAX_D);
+  EXPECT_DOUBLE_EQ(box1.Min().Z(), math::MAX_D);
 
-  EXPECT_NEAR(box1.Max().X(), 0, 1e-6);
-  EXPECT_NEAR(box1.Max().Y(), 0, 1e-6);
-  EXPECT_NEAR(box1.Max().Z(), 0, 1e-6);
+  EXPECT_DOUBLE_EQ(box1.Max().X(), math::LOW_D);
+  EXPECT_DOUBLE_EQ(box1.Max().Y(), math::LOW_D);
+  EXPECT_DOUBLE_EQ(box1.Max().Z(), math::LOW_D);
+}
+
+/////////////////////////////////////////////////
+TEST(BoxTest, DefaultConstructor)
+{
+  math::Box defaultBox1, defaultBox2;
+
+  EXPECT_DOUBLE_EQ(defaultBox1.Size().X(), 0);
+  EXPECT_DOUBLE_EQ(defaultBox1.Size().Y(), 0);
+  EXPECT_DOUBLE_EQ(defaultBox1.Size().Z(), 0);
+
+  EXPECT_DOUBLE_EQ(defaultBox1.XLength(), 0);
+  EXPECT_DOUBLE_EQ(defaultBox1.YLength(), 0);
+  EXPECT_DOUBLE_EQ(defaultBox1.ZLength(), 0);
+
+  EXPECT_DOUBLE_EQ(defaultBox1.Center().X(), 0);
+  EXPECT_DOUBLE_EQ(defaultBox1.Center().Y(), 0);
+  EXPECT_DOUBLE_EQ(defaultBox1.Center().Z(), 0);
+
+  EXPECT_FALSE(defaultBox1.Intersects(defaultBox2));
+
+  EXPECT_FALSE(defaultBox1.Intersects(math::Box(
+          math::Vector3d(0, 0, 0), math::Vector3d(1, 1, 1))));
+
+  EXPECT_FALSE(defaultBox2.Contains(math::Vector3d::Zero));
 }
 
 /////////////////////////////////////////////////
@@ -154,22 +214,22 @@ TEST(BoxTest, PlusEmpty)
   math::Box box2;
 
   box1 += box2;
-  EXPECT_NEAR(box1.Min().X(), 0, 1e-6);
-  EXPECT_NEAR(box1.Min().Y(), 0, 1e-6);
-  EXPECT_NEAR(box1.Min().Z(), 0, 1e-6);
+  EXPECT_DOUBLE_EQ(box1.Min().X(), math::MAX_D);
+  EXPECT_DOUBLE_EQ(box1.Min().Y(), math::MAX_D);
+  EXPECT_DOUBLE_EQ(box1.Min().Z(), math::MAX_D);
 
-  EXPECT_NEAR(box1.Max().X(), 0, 1e-6);
-  EXPECT_NEAR(box1.Max().Y(), 0, 1e-6);
-  EXPECT_NEAR(box1.Max().Z(), 0, 1e-6);
+  EXPECT_DOUBLE_EQ(box1.Max().X(), math::LOW_D);
+  EXPECT_DOUBLE_EQ(box1.Max().Y(), math::LOW_D);
+  EXPECT_DOUBLE_EQ(box1.Max().Z(), math::LOW_D);
 
   math::Box box3 = box2 + box1;
-  EXPECT_NEAR(box3.Min().X(), 0, 1e-6);
-  EXPECT_NEAR(box3.Min().Y(), 0, 1e-6);
-  EXPECT_NEAR(box3.Min().Z(), 0, 1e-6);
+  EXPECT_DOUBLE_EQ(box3.Min().X(), math::MAX_D);
+  EXPECT_DOUBLE_EQ(box3.Min().Y(), math::MAX_D);
+  EXPECT_DOUBLE_EQ(box3.Min().Z(), math::MAX_D);
 
-  EXPECT_NEAR(box3.Max().X(), 0, 1e-6);
-  EXPECT_NEAR(box3.Max().Y(), 0, 1e-6);
-  EXPECT_NEAR(box3.Max().Z(), 0, 1e-6);
+  EXPECT_DOUBLE_EQ(box3.Max().X(), math::LOW_D);
+  EXPECT_DOUBLE_EQ(box3.Max().Y(), math::LOW_D);
+  EXPECT_DOUBLE_EQ(box3.Max().Z(), math::LOW_D);
 }
 
 /////////////////////////////////////////////////